requirements.txt
streamlit==1.39.0
google-generativeai==0.8.5
langchain==0.3.7
langchain-community==0.3.7
langchain-text-splitters==0.3.2
faiss-cpu==1.8.0.post1
sentence-transformers==3.2.1
youtube-transcript-api==0.6.2
yt-dlp==2025.10.22
python-dotenv==1.0.1
